buggy-script
Initially, I thought there was an issue with one of the arguments for the writeFileSync method. The debugger was noting that path was supposed to be "string", "url", or "buffer" value type. After re-reading the issue, I then realized it is because in order to separate fileName from contentArr, the code was separating via " ". I then looked at the solution to confirm that all we had to do was join the contentArr at " ".